requirements.txt
spacy
python-dateutil